Some general fixes and improvements
More robust builder that does not require that the hook provides default value.
Moved babel configuration to a "global" one so that can be reused by other loaders, like isparta.
Fixed a small issue in the stats generation after allowing to change the built resource name.
Made sure that the build-webpack action runs always, this makes it possible to call it from other places like when running tests.